# docs/swagger.yaml
definitions:
  reqmodel.CreateDiscount:
    properties:
      end_date:
        type: string
      name:
        type: string
      start_date:
        type: string
      type:
        type: string
      value:
        type: number
    required:
    - end_date
    - name
    - start_date
    - type
    - value
    type: object
...
  reqmodel.CreateProductDiscountApplied:
    properties:
      customer_id:
        type: string
      customer_name:
        type: string
      discount_product_target_id:
        type: string
      transaction_date:
        type: string
    required:
    - discount_product_target_id
    type: object
...
  reqmodel.UpdateDiscount:
    properties:
      end_date:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      start_date:
        type: string
      type:
        type: string
      value:
        type: number
    type: object
...
  resmodel.DataResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  resmodel.DatasResponse:
    properties:
      datas:
        items: {}
        type: array
      message:
        type: string
      pagination: {}
      success:
        type: boolean
    type: object
  resmodel.NoDataResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
  utils.DataValidationError:
    properties:
      field:
        example: email
        type: string
      message:
        example: Invalid email address
        type: string
    type: object
  utils.RequestError:
    properties:
      fields:
        items:
          $ref: '#/definitions/utils.DataValidationError'
        type: array
      message:
        example: Invalid email address
        type: string
      status_code:
        example: 422
        type: integer
    type: object
info:
  contact:
    email: lifelinejar@mail.com
    name: API Support
  description: Service Discount Rest API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Service Discount
  version: "1.0"
paths:
  /ndi/discount:
    get:
      description: Get all discounts
      parameters:
      - description: Search keyword
        in: query
        name: search
        type: string
      - description: Page Number
        in: query
        name: page
        type: integer
      - description: Page Size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Discounts retrieved successfully
          schema:
            $ref: '#/definitions/resmodel.DatasResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.RequestError'
        "404":
          description: Discounts not found
          schema:
            $ref: '#/definitions/utils.RequestError'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/utils.RequestError'
      security:
      - ApiKeyAuth: []
      summary: List Discount
      tags:
      - Discount
    post:
      consumes:
      - application/json
      description: Create a new discount
      parameters:
      - description: User ID to check
        in: header
        name: X-Auth-User-Id
        required: true
        type: string
      - description: Create discount
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/reqmodel.CreateDiscount'
      produces:
      - application/json
      responses:
        "201":
          description: Discount created
          schema:
            $ref: '#/definitions/resmodel.NoDataResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.RequestError'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/utils.RequestError'
      security:
      - ApiKeyAuth: []
      summary: Create Discount
      tags:
      - Discount
...
  /ndi/discount/{id_unik}:
    delete:
      consumes:
      - application/json
      description: Delete a discount
      parameters:
      - description: User ID to check
        in: header
        name: X-Auth-User-Id
        required: true
        type: string
      - description: Discount ID
        in: path
        name: id_unik
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Discount deleted
          schema:
            $ref: '#/definitions/resmodel.NoDataResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.RequestError'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/utils.RequestError'
      security:
      - ApiKeyAuth: []
      summary: Delete Discount
      tags:
      - Discount
    get:
      description: Get a discount by its ID
      parameters:
      - description: Discount ID
        in: path
        name: id_unik
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Discount retrieved successfully
          schema:
            $ref: '#/definitions/resmodel.DataResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.RequestError'
        "404":
          description: Discount not found
          schema:
            $ref: '#/definitions/utils.RequestError'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/utils.RequestError'
      security:
      - ApiKeyAuth: []
      summary: Get Discount By ID
      tags:
      - Discount
    put:
      consumes:
      - application/json
      description: Update an existing discount
      parameters:
      - description: User ID to check
        in: header
        name: X-Auth-User-Id
        required: true
        type: string
      - description: Discount ID
        in: path
        name: id_unik
        required: true
        type: string
      - description: Update discount
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/reqmodel.UpdateDiscount'
      produces:
      - application/json
      responses:
        "200":
          description: Discount updated
          schema:
            $ref: '#/definitions/resmodel.NoDataResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.RequestError'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/utils.RequestError'
      security:
      - ApiKeyAuth: []
      summary: Update Discount
      tags:
      - Discount
  /ndi/product-discount-applied:
    get:
      description: Get all product discount applied
      parameters:
      - description: Search keyword
        in: query
        name: search
        type: string
      - description: Page Number
        in: query
        name: page
        type: integer
      - description: Page Size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Product discount applied retrieved successfully
          schema:
            $ref: '#/definitions/resmodel.DatasResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.RequestError'
        "404":
          description: Product discount applied not found
          schema:
            $ref: '#/definitions/utils.RequestError'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/utils.RequestError'
      security:
      - ApiKeyAuth: []
      summary: List Product Discount Applied
      tags:
      - Product Discount Applied
    post:
      consumes:
      - application/json
      description: Create a new product discount applied
      parameters:
      - description: User ID to check
        in: header
        name: X-Auth-User-Id
        required: true
        type: string
      - description: Create product discount applied
        in: body
        name: request
        required: true
        schema:
          items:
            $ref: '#/definitions/reqmodel.CreateProductDiscountApplied'
          type: array
      produces:
      - application/json
      responses:
        "201":
          description: Product discount applied created
          schema:
            $ref: '#/definitions/resmodel.NoDataResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.RequestError'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/utils.RequestError'
      security:
      - ApiKeyAuth: []
      summary: Create Product Discount Applied
      tags:
      - Product Discount Applied
  /ndi/product-discount-applied/{id_unik}:
    delete:
      consumes:
      - application/json
      description: Delete a product discount applied
      parameters:
      - description: User ID to check
        in: header
        name: X-Auth-User-Id
        required: true
        type: string
      - description: Product Discount Applied ID
        in: path
        name: id_unik
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Product discount applied deleted
          schema:
            $ref: '#/definitions/resmodel.NoDataResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.RequestError'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/utils.RequestError'
      security:
      - ApiKeyAuth: []
      summary: Delete Product Discount Applied
      tags:
      - Product Discount Applied
  /ndi/product-discount-applied/by-discount/{id_unik}:
    get:
      description: Get a list of product discount applied by its discount ID
      parameters:
      - description: Product Discount ID
        in: path
        name: id_unik
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Product discount applied retrieved successfully
          schema:
            $ref: '#/definitions/resmodel.DatasResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.RequestError'
        "404":
          description: Product discount applied not found
          schema:
            $ref: '#/definitions/utils.RequestError'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/utils.RequestError'
      security:
      - ApiKeyAuth: []
      summary: List Product Discount Applied By Discount ID
      tags:
      - Product Discount Applied
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"